# take_while1: rely on take_while entirely for its implementation